pyproject.toml
[tool.poetry]
name = "django-kafka-consumer"
version = "2.0.0"
description = ""
authors = ["IIIT <github@iiit.pl>"]
readme = "README.md"
license = "BSD-2-Clause"
repository = "https://github.com/innovationinit/django-kafka-consumer"
packages = [
    { include = "kafka_consumer" },
]
include = [
    "README.md",
    "LICENSE",
    "kafka_consumer/locale/**",
]


[tool.poetry.dependencies]
python = ">=3.7,<3.10"
kafka-python = "2.0.2"
Django = ">=2.0,<3.0"
pytz = "^2021.3"

[tool.poetry.dev-dependencies]
pycrypto = "^2.6.1"
tox = "^3.24.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
